Notification restore revamp to use Worker now   * **Deleted Files:**     * `NotificationRestorer`     * `NotificationRestoreService`     * `RestoreJobService`     * `RestoreKickoffJobService`     * Services all removed from `AndroidManifest` as well since `Worker` acts as a replacement   * `UpgradeReceiver` and `BootUpReceiver` now kickoff notification restoring through `OSNotificationRestoreWorkManager`
Added boolean shouldDelay param to restore work manager
  * `OSNotificationRestoreWorkManager.beginEnqueueingWork` should have a `boolean shouldDelay` param
  * When app comes from boot or upgrade, add a 15 second delay to alleviate app doing to much work all at once

Fixed UnitTests related to app not being in foreground
  * The foreground issue is related to foreground handlers not firing now when app is backgrounded
  * Also fixed a network request assert in another UnitTest
  * `shouldGenerate2BasicGroupNotifications`
