+++ OneSignalSDK/onesignal/src/main/java/com/onesignal/NotificationBundleProcessor.java
@@ -93,8 +93,10 @@ class NotificationBundleProcessor {
                     && OneSignal.notValidOrDuplicated(context, jsonPayload))
                 return;
 
+            String osNotificationId = OSNotificationFormatHelper.getOSNotificationIdFromJson(jsonPayload);
             OSNotificationWorkManager.beginEnqueueingWork(
                     context,
+                    osNotificationId,
                     androidNotificationId,
                     jsonStrPayload,
                     isRestoring,
@@ -227,7 +229,7 @@ class NotificationBundleProcessor {
 
                 // Set expire_time
                 long sentTime = jsonPayload.optLong("google.sent_time", SystemClock.currentThreadTimeMillis()) / 1_000L;
-                int ttl = jsonPayload.optInt("google.ttl", NotificationRestorer.DEFAULT_TTL_IF_NOT_IN_PAYLOAD);
+                int ttl = jsonPayload.optInt("google.ttl", OSNotificationRestoreWorkManager.DEFAULT_TTL_IF_NOT_IN_PAYLOAD);
                 long expireTime = sentTime + ttl;
                 values.put(NotificationTable.COLUMN_NAME_EXPIRE_TIME, expireTime);
