+++ OneSignalSDK/unittest/src/test/java/com/onesignal/OneSignalPackagePrivateHelper.java
@@ -182,7 +182,7 @@ public class OneSignalPackagePrivateHelper {
 
    public static class NotificationTable extends OneSignalDbContract.NotificationTable { }
    public static class InAppMessageTable extends OneSignalDbContract.InAppMessageTable { }
-   public static class NotificationRestorer extends com.onesignal.NotificationRestorer { }
+   public static class OSNotificationRestoreWorkManager extends com.onesignal.OSNotificationRestoreWorkManager { }
     public static class OSNotificationGenerationJob extends com.onesignal.OSNotificationGenerationJob {
         OSNotificationGenerationJob(Context context) {
             super(context);
