+++ OneSignalSDK/unittest/src/test/java/com/test/onesignal/GenerateNotificationRunner.java
@@ -59,11 +59,10 @@ import com.onesignal.OSNotificationReceived;
 import com.onesignal.OneSignal;
 import com.onesignal.OneSignalNotificationManagerPackageHelper;
 import com.onesignal.OneSignalPackagePrivateHelper;
-import com.onesignal.OneSignalPackagePrivateHelper.NotificationRestorer;
+import com.onesignal.OneSignalPackagePrivateHelper.OSNotificationRestoreWorkManager;
 import com.onesignal.OneSignalPackagePrivateHelper.NotificationTable;
 import com.onesignal.OneSignalPackagePrivateHelper.TestOneSignalPrefs;
 import com.onesignal.OneSignalShadowPackageManager;
-import com.onesignal.RestoreJobService;
 import com.onesignal.ShadowBadgeCountUpdater;
 import com.onesignal.ShadowCustomTabsClient;
 import com.onesignal.ShadowCustomTabsSession;
@@ -185,7 +184,7 @@ public class GenerateNotificationRunner {
 
       NotificationManager notificationManager = OneSignalNotificationManagerPackageHelper.getNotificationManager(blankActivity);
       notificationManager.cancelAll();
-      NotificationRestorer.restored = false;
+      OSNotificationRestoreWorkManager.restored = false;
    }
 
    @AfterClass
@@ -257,7 +256,7 @@ public class GenerateNotificationRunner {
       threadAndTaskWait();
 
       // Restore notifs
-      NotificationRestorer.restore(blankActivity);
+      OSNotificationRestoreWorkManager.beginEnqueueingWork(blankActivity, false);
       threadAndTaskWait();
 
       // Assert that no restoration jobs were scheduled
@@ -304,7 +303,7 @@ public class GenerateNotificationRunner {
       Intent intent = Shadows.shadowOf(postedNotification.notif.contentIntent).getSavedIntent();
       NotificationOpenedProcessor_processFromContext(blankActivity, intent);
       threadAndTaskWait();
-      
+
       // Make sure we get a payload when it is opened.
       assertNotNull(lastOpenResult.notification.payload);
    }
@@ -344,7 +343,7 @@ public class GenerateNotificationRunner {
       Bundle bundle = getBaseNotifBundle("UUID0");
       NotificationBundleProcessor_ProcessFromFCMIntentService(blankActivity, bundle, null);
       threadAndTaskWait();
-      
+
       bundle = getBaseNotifBundle("UUID1");
       bundle.putString("grp", "test1");
       NotificationBundleProcessor_ProcessFromFCMIntentService(blankActivity, bundle, null);
@@ -602,12 +601,12 @@ public class GenerateNotificationRunner {
       bundle.putString("grp", "test1");
       NotificationBundleProcessor_ProcessFromFCMIntentService(blankActivity, bundle, null);
       threadAndTaskWait();
-   
+
       bundle = getBaseNotifBundle("UUID2");
       bundle.putString("grp", "test1");
       NotificationBundleProcessor_ProcessFromFCMIntentService(blankActivity, bundle, null);
       threadAndTaskWait();
-   
+
       bundle = getBaseNotifBundle("UUID3");
       bundle.putString("grp", "test1");
       NotificationBundleProcessor_ProcessFromFCMIntentService(blankActivity, bundle, null);
@@ -618,7 +617,7 @@ public class GenerateNotificationRunner {
       
       // Test - 3 notifis + 1 summary
       assertEquals(4, postedNotifs.size());
-      
+
       // Test - First notification should be the summary
       PostedNotification postedSummaryNotification = postedNotifsIterator.next().getValue();
       assertEquals("3 new messages", postedSummaryNotification.getShadow().getContentText());
@@ -627,7 +626,7 @@ public class GenerateNotificationRunner {
       // Setup - Let's cancel a child notification.
       PostedNotification postedNotification = postedNotifsIterator.next().getValue();
       OneSignal.cancelNotification(postedNotification.id);
-   
+
       // Test - It should update summary text to say 2 notifications
       postedNotifs = ShadowRoboNotificationManager.notifications;
       assertEquals(3, postedNotifs.size());       // 2 notifis + 1 summary
@@ -656,7 +655,7 @@ public class GenerateNotificationRunner {
       
       // Setup - Let's cancel our 3rd and last child notification.
       OneSignal.cancelNotification(postedNotification.id);
-   
+
       // Test - No more notifications! :)
       postedNotifs = ShadowRoboNotificationManager.notifications;
       assertEquals(0, postedNotifs.size());
@@ -710,20 +709,22 @@ public class GenerateNotificationRunner {
 
    @Test
    @Config(shadows = { ShadowGenerateNotification.class })
-   public void shouldSetBadgesWhenRestoringNotifications() {
+   public void shouldSetBadgesWhenRestoringNotifications() throws Exception {
       NotificationBundleProcessor_ProcessFromFCMIntentService(blankActivity, getBaseNotifBundle(), null);
       ShadowBadgeCountUpdater.lastCount = 0;
 
-      NotificationRestorer.restore(blankActivity);
+      OSNotificationRestoreWorkManager.beginEnqueueingWork(blankActivity, false);
+      threadAndTaskWait();
 
       assertEquals(1, ShadowBadgeCountUpdater.lastCount);
    }
 
-   @Test public void shouldNotRestoreNotificationsIfPermissionIsDisabled() {
+   @Test public void shouldNotRestoreNotificationsIfPermissionIsDisabled() throws Exception {
       NotificationBundleProcessor_ProcessFromFCMIntentService(blankActivity, getBaseNotifBundle(), null);
 
       ShadowNotificationManagerCompat.enabled = false;
-      NotificationRestorer.restore(blankActivity);
+      OSNotificationRestoreWorkManager.beginEnqueueingWork(blankActivity, false);
+      threadAndTaskWait();
 
       assertNull(Shadows.shadowOf(blankActivity).getNextStartedService());
    }
@@ -782,7 +783,7 @@ public class GenerateNotificationRunner {
       bundle2.putBoolean("is_restoring", true);
       NotificationBundleProcessor_ProcessFromFCMIntentService_NoWrap(blankActivity, bundle2, null);
       threadAndTaskWait();
-      
+
       // Test - Restored notifications display exactly the same as they did when received.
       postedNotifs = ShadowRoboNotificationManager.notifications;
       postedNotifsIterator = postedNotifs.entrySet().iterator();
@@ -854,38 +855,40 @@ public class GenerateNotificationRunner {
 
    @Test
    @Config(shadows = { ShadowGenerateNotification.class })
-   public void shouldRestoreNotifications() {
-      NotificationRestorer.restore(blankActivity); NotificationRestorer.restored = false;
+   public void shouldRestoreNotifications() throws Exception {
+      restoreNotifications();
+      threadAndTaskWait();
+
+      assertEquals(0, ShadowBadgeCountUpdater.lastCount);
 
       NotificationBundleProcessor_ProcessFromFCMIntentService(blankActivity, getBaseNotifBundle(), null);
 
-      NotificationRestorer.restore(blankActivity); NotificationRestorer.restored = false;
-      Intent intent = Shadows.shadowOf(blankActivity).getNextStartedService();
-      assertEquals(RestoreJobService.class.getName(), intent.getComponent().getClassName());
+      restoreNotifications();
+      threadAndTaskWait();
+
+      assertEquals(1, ShadowBadgeCountUpdater.lastCount);
+
+//      Intent intent = Shadows.shadowOf(blankActivity).getNextStartedService();
+//      assertEquals(RestoreJobService.class.getName(), intent.getComponent().getClassName());
 
       // Go forward 1 week
       // Note: Does not effect the SQL function strftime
       advanceSystemTimeBy(604_801);
 
       // Restorer should not fire service since the notification is over 1 week old.
-      NotificationRestorer.restore(blankActivity); NotificationRestorer.restored = false;
-      assertNull(Shadows.shadowOf(blankActivity).getNextStartedService());
-   }
+      restoreNotifications();
+      threadAndTaskWait();
+//      assertNull(Shadows.shadowOf(blankActivity).getNextStartedService());
 
-   private void assertRestoreRan() {
-      Intent intent = Shadows.shadowOf(blankActivity).getNextStartedService();
-      assertEquals(RestoreJobService.class.getName(), intent.getComponent().getClassName());
-   }
-   private void assertRestoreDidNotRun() {
-      assertNull(Shadows.shadowOf(blankActivity).getNextStartedService());
+      assertEquals(0, ShadowBadgeCountUpdater.lastCount);
    }
 
    private void restoreNotifications() {
-      NotificationRestorer.restored = false;
-      NotificationRestorer.restore(blankActivity);
+      OSNotificationRestoreWorkManager.restored = false;
+      OSNotificationRestoreWorkManager.beginEnqueueingWork(blankActivity, false);
    }
 
-   private void helperShouldRestoreNotificationsPastExpireTime(boolean should) {
+   private void helperShouldRestoreNotificationsPastExpireTime(boolean should) throws Exception {
       long ttl = 60L;
       Bundle bundle = getBaseNotifBundle();
       bundle.putLong("google.sent_time", System.currentTimeMillis());
@@ -893,26 +896,28 @@ public class GenerateNotificationRunner {
       NotificationBundleProcessor_ProcessFromFCMIntentService(blankActivity, bundle, null);
 
       restoreNotifications();
-      assertRestoreRan();
+      threadAndTaskWait();
+
+      assertEquals(1, ShadowBadgeCountUpdater.lastCount);
 
       // Go forward just past the TTL of the notification
       advanceSystemTimeBy(ttl + 1);
       restoreNotifications();
       if (should)
-         assertRestoreRan();
+         assertEquals(1, ShadowBadgeCountUpdater.lastCount);
       else
-         assertRestoreDidNotRun();
+         assertEquals(0, ShadowBadgeCountUpdater.lastCount);
    }
 
    @Test
    @Config(shadows = { ShadowGenerateNotification.class })
-   public void doNotRestoreNotificationsPastExpireTime() {
+   public void doNotRestoreNotificationsPastExpireTime() throws Exception {
       helperShouldRestoreNotificationsPastExpireTime(false);
    }
 
    @Test
    @Config(shadows = { ShadowGenerateNotification.class })
-   public void restoreNotificationsPastExpireTimeIfSettingIsDisabled() {
+   public void restoreNotificationsPastExpireTimeIfSettingIsDisabled() throws Exception {
       TestOneSignalPrefs.saveBool(TestOneSignalPrefs.PREFS_ONESIGNAL, TestOneSignalPrefs.PREFS_OS_RESTORE_TTL_FILTER, false);
       helperShouldRestoreNotificationsPastExpireTime(true);
    }
@@ -1035,7 +1040,7 @@ public class GenerateNotificationRunner {
 
       FCMBroadcastReceiver_onReceived_withIntent(blankActivity, intent);
       threadAndTaskWait();
-      
+
       // Normal notifications should be generated right from the BroadcastReceiver
       //   without creating a service.
       assertNull(Shadows.shadowOf(blankActivity).getNextStartedService());
@@ -1521,6 +1526,9 @@ public class GenerateNotificationRunner {
       });
       threadAndTaskWait();
 
+      blankActivityController.resume();
+      threadAndTaskWait();
+
       // 3. Receive a notification
       FCMBroadcastReceiver_processBundle(blankActivity, getBaseNotifBundle());
       threadAndTaskWait();
@@ -1552,6 +1560,9 @@ public class GenerateNotificationRunner {
             notifJob.complete();
          }
       });
+      threadAndTaskWait();
+
+      blankActivityController.resume();
       threadAndTaskWait();
 
       // 3. Receive a notification
@@ -1602,6 +1613,9 @@ public class GenerateNotificationRunner {
       });
       threadAndTaskWait();
 
+      blankActivityController.resume();
+      threadAndTaskWait();
+
       // 3. Receive a notification
       FCMBroadcastReceiver_processBundle(blankActivity, getBaseNotifBundle());
       threadAndTaskWait();
@@ -1634,6 +1648,9 @@ public class GenerateNotificationRunner {
             notifJob.complete();
          }
       });
+      threadAndTaskWait();
+
+      blankActivityController.resume();
       threadAndTaskWait();
 
       // 3. Receive a notification
@@ -1684,6 +1701,9 @@ public class GenerateNotificationRunner {
       });
       threadAndTaskWait();
 
+      blankActivityController.resume();
+      threadAndTaskWait();
+
       // 3. Receive a notification
       FCMBroadcastReceiver_processBundle(blankActivity, getBaseNotifBundle());
       threadAndTaskWait();
@@ -1707,6 +1727,9 @@ public class GenerateNotificationRunner {
       // 2. Init OneSignal
       OneSignal.setAppId("b2f7f966-d8cc-11e4-bed1-df8f05be55ba");
       OneSignal.setAppContext(blankActivity);
+      threadAndTaskWait();
+
+      blankActivityController.resume();
       threadAndTaskWait();
 
       // 3. Receive a notification
@@ -1792,6 +1815,9 @@ public class GenerateNotificationRunner {
       });
       threadAndTaskWait();
 
+      blankActivityController.resume();
+      threadAndTaskWait();
+
       // 3. Receive a notification
       FCMBroadcastReceiver_processBundle(blankActivity, getBaseNotifBundle());
       threadAndTaskWait();
@@ -1815,6 +1841,9 @@ public class GenerateNotificationRunner {
       // 2. Init OneSignal
       OneSignal.setAppId("b2f7f966-d8cc-11e4-bed1-df8f05be55ba");
       OneSignal.setAppContext(blankActivity);
+      threadAndTaskWait();
+
+      blankActivityController.resume();
       threadAndTaskWait();
 
       // 3. Receive a notification
@@ -1864,6 +1893,9 @@ public class GenerateNotificationRunner {
       });
       threadAndTaskWait();
 
+      blankActivityController.resume();
+      threadAndTaskWait();
+
       // 3. Receive a notification
       FCMBroadcastReceiver_processBundle(blankActivity, getBaseNotifBundle());
       threadAndTaskWait();
@@ -1890,6 +1922,9 @@ public class GenerateNotificationRunner {
       // 2. Init OneSignal
       OneSignal.setAppId("b2f7f966-d8cc-11e4-bed1-df8f05be55ba");
       OneSignal.setAppContext(blankActivity);
+      threadAndTaskWait();
+
+      blankActivityController.resume();
       threadAndTaskWait();
 
       // 3. Receive a notification
@@ -1926,6 +1961,9 @@ public class GenerateNotificationRunner {
             // Complete is not called, so we rely on 30 second timeout until bubbling by default out of the AppNotificationWillShowInForegroundHandler
          }
       });
+      threadAndTaskWait();
+
+      blankActivityController.resume();
       threadAndTaskWait();
 
       // 3. Receive a notification
@@ -1987,6 +2025,9 @@ public class GenerateNotificationRunner {
             notifJob.complete();
          }
       });
+
+      blankActivityController.resume();
+      threadAndTaskWait();
 
       // 3. Receive a notification
       // Setup - Display a single notification with a grouped.
@@ -2067,6 +2108,9 @@ public class GenerateNotificationRunner {
             }, 35_000L);
          }
       });
+      threadAndTaskWait();
+
+      blankActivityController.resume();
       threadAndTaskWait();
 
       // 3. Receive a notification
